packets (#73)
<!-- OBSERVAÇÃO: Qualquer conteúdo dentro dessas marcas será ocultado na
pré-visualização do Pull Request. -->

* **Problema(s) Tratado(s)**:

<!--
Por favor, especifique os
[problemas](https://help.github.com/articles/autolinked-references-and-urls/#issues-and-pull-requests)
do brHades que este pull request aborda.
Se ainda não houver um problema, crie um
[aqui](https://github.com/AoShinRo/brHades/issues/new) primeiro e depois
vincule seu pull request a esse problema!
-->

* **Modo do Servidor**:

<!-- Em qual modo este pull request se aplica: Pré-Renewal, Renewal ou
Ambos? -->

* **Descrição do Pull Request**:

<!-- Descreva como este pull request resolverá os problemas listados
acima. -->

+++ src/map/skill.hpp
@@ -373,7 +373,7 @@ struct skill_unit {
 
 /// Skill unit group
 struct s_skill_unit_group {
-	int src_id; /// Caster ID/RID, if player is account_id
+	uint32 src_id; /// Caster ID/RID, if player is account_id
 	int party_id; /// Party ID
 	int guild_id; /// Guild ID
 	int bg_id; /// Battleground ID
@@ -387,7 +387,7 @@ struct s_skill_unit_group {
 		skill_lv; /// Skill level
 	int val1, val2, val3; /// Values
 	char *valstr; /// String value, used for HT_TALKIEBOX & RG_GRAFFITI
-	int unit_id; /// Unit ID (for client effect)
+	uint16 unit_id; /// Unit ID (for client effect)
 	int group_id; /// Skill Group ID
 	int link_group_id; /// Linked group that should be deleted if this one is deleted
 	int unit_count, /// Number of unit at this group
